requirements.txt
# Core Deep Learning
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.3.0
albumentations>=1.3.0

# Data Science & Analysis
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Application
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
openpyxl>=3.0.0

# COCO Tools
pycocotools>=2.0.0

# Optional: Experiment Tracking
# wandb>=0.15.0
# tensorboard>=2.10.0

# Optional: Model Optimization  
# onnx>=1.12.0
# onnxruntime>=1.12.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0